requirements.txt
Flask
Flask-CORS
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
openai 